refactor(touchcapture)!: move to utils directory
